executor/sql_executor_test.go
package executor

import (
	"context"
	"os"
	"testing"
	"time"

	"github.com/iyuangang/oracle-sql-runner/config"
)

func TestSQLExecutor(t *testing.T) {
	// 加载测试配置
	cfg := &config.DatabaseConfig{
		Name:          "test",
		User:          os.Getenv("TEST_DB_USER"),
		Password:      os.Getenv("TEST_DB_PASSWORD"),
		Host:          os.Getenv("TEST_DB_HOST"),
		Port:          1521,
		Service:       os.Getenv("TEST_DB_SERVICE"),
		AutoCommit:    true,
		MaxRetries:    3,
		Timeout:       30 * time.Second,
		EnableDBMSOut: true,
	}

	execCfg := &config.ExecutionConfig{
		ParallelDegree: 4,
		BatchSize:      1000,
		MaxFileSize:    104857600,
		RetryInterval:  5 * time.Second,
	}

	// 创建执行器
	exec, err := NewSQLExecutor(cfg, execCfg)
	if err != nil {
		t.Fatalf("创建执行器失败: %v", err)
	}
	defer exec.Close()

	// 测试用例
	tests := []struct {
		name    string
		sql     string
		wantErr bool
	}{
		{
			name: "Simple SELECT",
			sql:  "SELECT 1 FROM DUAL",
		},
		{
			name: "Create Table",
			sql: `CREATE TABLE test_table (
                id NUMBER PRIMARY KEY,
                name VARCHAR2(100)
            )`,
		},
		{
			name: "Insert Data",
			sql:  "INSERT INTO test_table VALUES (1, 'test')",
		},
		{
			name: "PL/SQL Block",
			sql: `
            BEGIN
                DBMS_OUTPUT.PUT_LINE('Hello, World!');
            END;
            `,
		},
		{
			name:    "Invalid SQL",
			sql:     "SELECT * FROM non_existent_table",
			wantErr: true,
		},
	}

	// 执行测试
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			conn, err := exec.pool.Acquire(context.Background())
			if err != nil {
				t.Fatalf("获取连接失败: %v", err)
			}
			defer exec.pool.Release(conn)

			err = exec.executeStatement(conn, tt.sql)
			if (err != nil) != tt.wantErr {
				t.Errorf("executeStatement() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestParallelExecution(t *testing.T) {
	// 测试并行执行
	// ... 实现并行执行测试 ...
}

func TestErrorHandling(t *testing.T) {
	// 测试错误处理
	// ... 实现错误处理测试 ...
}

func TestTransactionManagement(t *testing.T) {
	// 测试事务管理
	// ... 实现事务管理测试 ...
}
